Redo all the syntax contains lines
When embedding a syntax file in another with :syn-include, all uses of
contains=TOP,foo are broken. Since we depended on that a lot, this made
embedding not work very well. As such, rewrite the syntax file to not
use that anymore.

Note: I think the new contains expressions are equivalent to the old,
but it's always possible that I may have screwed something up in the
manual translation. Let me know if anything highlights wrong.