Fix the handling of buffer state after :SwiftEmit*
The s:WithPath() function was created for :SwiftRun, and when reused for
:SwiftEmit* it didn't properly restore state on the original buffer,
instead restoring it on the current buffer. This meant that, for
example, using a :SwiftEmit* command on a scratch buffer would
incorrectly leave the buffer marked as unmodified.

In the process, get rid of the use of :lcd and rely instead on telling
swiftc what output directory to use. This is necessary because it
appears to be effectively impossible to restore the original cwd state
after using :lcd and running arbitrary code.